Plates: Fixed a crash loading the package's test suite in DevTools Test Manager
