+++ packages/Plates/CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## v.next
 
-* Analyses: break circular import between `base-analysis.ts` and `package.ts` to fix TDZ error `Cannot access 'AnalysisBase' before initialization` when the test bundle loads.
+* Plates: Fixed a crash loading the package's test suite in DevTools Test Manager.
