+++ packages/Plates/src/plate/analyses/base-analysis.ts
@@ -11,6 +11,7 @@ import {
   saveAnalysisResult, saveAnalysisRunParameter
 } from '../../plates/plates-crud';
 import {AnalysisRequiredFields} from '../../plates/views/components/analysis-mapping/analysis-mapping-panel';
+import {_package} from '../../package';
 import {PlateWidget} from '../plate-widget/plate-widget';
 
 export interface IAnalysisProperty {
@@ -153,7 +154,7 @@ export abstract class AnalysisBase implements IPlateAnalysis {
         grok.shell.info(`Saved ${this.friendlyName} results for plate ${plate.barcode}.`);
       } catch (e: any) {
         const errorMessage = `Failed to save ${this.friendlyName} results. Reason: ${e.message}`;
-        console.error(errorMessage);
+        _package.logger.debug(errorMessage);
         if (runId) {
           try {
             await grok.data.db.query('Plates:Plts', `DELETE FROM plts.analysis_runs WHERE id = ${runId};`);
